PM-2.02 test: adding tests, refactoring test refactor: production code fix: problems with movieItem list

+++ app/src/main/java/com/example/popularmovie/app/data/MovieDbHelper.java
@@ -3,6 +3,7 @@ package com.example.popularmovie.app.data;
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
+import android.util.Log;
 
 import com.example.popularmovie.app.data.MovieContract.MovieEntry;
 
@@ -12,9 +13,11 @@ import com.example.popularmovie.app.data.MovieContract.MovieEntry;
  */
 public class MovieDbHelper extends SQLiteOpenHelper {
 
+    private static final String LOG_TAG = MovieDbHelper.class.getSimpleName();
+
     private static final int DATABASE_VERSION = 1;
 
-    static final String DATABASE_NAME = "movie.db";
+    static final String DATABASE_NAME = "movieDB.db";
 
 
     public MovieDbHelper(Context context) {
@@ -27,19 +30,24 @@ public class MovieDbHelper extends SQLiteOpenHelper {
                 "CREATE TABLE  " +
                         MovieEntry.TABLE_NAME +
                         " (" +
-                        MovieEntry.COLUMN_ID + " INTEGER NOT NULL " +
-                        MovieEntry.COLUMN_TITLE + " TEXT NOT NULL " +
-                        MovieEntry.COLUMN_RELEASE_DATE + " INTEGER NOT NULL " +
-                        MovieEntry.COLUMN_POSTER_URL + " TEXT NOT NULL " +
-                        MovieEntry.COLUMN_OVERVIEW +  " TEXT NOT NULL " +
-                        MovieEntry.COLUMN_RATING + " REAL NOT NULL " +
-                        MovieEntry.COLUMN_RUNTIME + " INTEGER NOT NULL" +
+                        MovieEntry.COLUMN_ID + " INTEGER PRIMARY KEY, " +
+                        MovieEntry.COLUMN_TITLE + " TEXT NOT NULL, " +
+                        MovieEntry.COLUMN_RELEASE_DATE + " TEXT NOT NULL, " +
+                        MovieEntry.COLUMN_POSTER + " TEXT NOT NULL, " +
+                        MovieEntry.COLUMN_OVERVIEW + " TEXT NOT NULL, " +
+                        MovieEntry.COLUMN_RATING + " REAL NOT NULL, " +
+                        MovieEntry.COLUMN_RUNTIME + " INTEGER NOT NULL " +
                         ");";
+        Log.d(LOG_TAG, SQL_CREATE_MOVIE_TABLE);
+        db.execSQL(SQL_CREATE_MOVIE_TABLE);
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + MovieEntry.TABLE_NAME);
-        onCreate(db);
+        if (newVersion > oldVersion) {
+            Log.w(LOG_TAG, " Old database version is : "+oldVersion+ " New database version is:"+newVersion);
+            db.execSQL("DROP TABLE IF EXISTS " + MovieEntry.TABLE_NAME);
+            onCreate(db);
+        }
     }
 }
